Move App and AdminPage/hooks
