+++ client/src/App/Splash.tsx
