+++ client/src/components/AdminPage/EventAdmin/Reports.tsx
@@ -1,5 +1,4 @@
 import React from 'react';
-// import { useEvent } from '../../../hooks';
 
 interface Props {
   event: ApiEvent,
